Move debug function to root
